+++ examples/passthrough/passthrough.go
@@ -21,8 +21,8 @@ import (
 	"path/filepath"
 	"syscall"
 
-	"github.com/AmericanBinary/cgofuse/examples/shared"
-	"github.com/AmericanBinary/cgofuse/fuse"
+	"github.com/americanbinary/cgofuse/examples/shared"
+	"github.com/americanbinary/cgofuse/fuse"
 )
 
 func trace(vals ...interface{}) func(vals ...interface{}) {